posts can now be nested in folders

+++ index.js
@@ -12,7 +12,7 @@ const
     Handlebars = require('handlebars'),
     layouts = require('handlebars-layouts'),
     sitemap = require('sitemap'),
-    converter = new showdown.Converter(),
+    converter = new showdown.Converter({ tables : true }),
     pages = {},
     posts = {},
     pageSize = parseInt(process.env.pageSize ? process.env.pageSize : '3'),
@@ -31,10 +31,18 @@ fs.ensureDirSync(archiveFolder);
 
 Handlebars.registerHelper(layouts(Handlebars));
 
+// to make it easier to keep static content like css, JS etc in /web for live updating, we reserve web/static for that content. Everything outside this folder will be
+// nuked during cleanup.
+if (fs.existsSync('./posts/static')){
+    console.log('ERROR : "posts/static" is a reserved folder name and may collide with protected static content in web/static. Please remove this folder.');
+    process.exit(1);
+}
+
 // clean up rendered files
-const htmlFiles = glob.sync(path.join(outFolder, '**/*.html'));
+const htmlFiles = glob.sync(path.join(outFolder, '**/*.*'), { ignore : path.join(outFolder, 'static/**') });
 for (let file of htmlFiles)
     fs.unlinkSync(file);
+
 
 // register partial templates
 const partialPaths = glob.sync('./templates/partials/**/*.hbs'); 
@@ -61,6 +69,7 @@ for (let helperPath of helperPaths){
 }
 
 
+
 /**
  * extract models from the top of markdown files
  * required fields for model :
@@ -69,12 +78,16 @@ for (let helperPath of helperPaths){
  * Note as well that the field "markup" is reserved, and will end up containing the markup rendered from markdown content
  */ 
 const markdownPaths = glob.sync('./posts/**/*.md'); 
+console.log(`found ${markdownPaths.length} articles to render.`);
+
 for (const markdownPath of markdownPaths){
+    
+    // post name is its entire path under ./posts, minus extension, for example "./posts/foo/bar.md" becomes "/foo/bar". The
+    // leading / is important for relative HTML mapping to work
     let post = {},
         content = fs.readFileSync(markdownPath, 'utf8'),
-        name = path.basename(markdownPath).match(/(.*).md/).pop();
-
-    let lines = content.split('\n');
+        name = markdownPath.substring('./posts'.length).match(/(.*).md/).pop(),
+        lines = content.split('\n'); 
 
     // find the position of the dividing line between post data and markup. dividing line is 3 or more dashes, egs "---"
     let dividerLineCount = null;
@@ -112,30 +125,48 @@ for (const markdownPath of markdownPaths){
     // tags are optional, if none are defined, create empty list. tags must be entered as comma-separated
     // list, but are converted to array
     post.tags = post.tags || '';
-    post.tags = post.tags.split(',').filter((tag)=>{ return tag.length > 0;});
-    post.tags = post.tags.map((tag)=>{return tag.trim()});
 
+    // remove empty items
+    post.tags = post.tags.split(',').filter((tag)=>{ return tag.length > 0;}); 
+
+    // remove whitespace around each tag
+    post.tags = post.tags.map((tag)=>{return tag.trim()}); 
+
+    // force convert date string to date object. Any valid JS datestring is valid.
     post.date = new Date(post.date);
     if (isNaN(post.date)){
         console.error(`WARNING : post ${name} has an invalid or missing date, it won't be published.`);
         continue;
     }
 
-    // post url is locked to the filename of its markdown file
+    // post url is locked to the relative path+name of its markdown file
     post.url = `${name}`;
 
-    // convert markdown to markup
+    // convert markdown (everything after model divider line) into HTML
     post.markup = lines.slice(dividerLineCount + 1).join('\n');
     post.markup = converter.makeHtml(post.markup);
-    post.keywords = post.tags.join(',');
 
+    post.keywords = post.tags.join(',');
     posts[name] = post;
 }
 
+
+// index.html is reserved, warn user if post called "index" is found
 if (posts.index){
     console.error('WARNING - you have a post called "index", this name is reserved and will always be overwritten');
 }
-    
+
+
+// sync static files from ./posts folder - this is options
+if (process.env.syncStaticPostFiles){
+    const syncFiles = glob.sync('./posts/' + '**/*.{png,gif,jpg}'); 
+    for (const syncFile of syncFiles){
+        const targetFile = syncFile.substring('./posts'.length);
+        fs.copySync(syncFile, path.join(outFolder, targetFile));
+    }
+}
+
+
 
 let tags = {},
     menuItems = [],
@@ -173,6 +204,7 @@ for (let i = 0; i < allPosts.length; i ++){
 
     let rendered = pages.post({ previousPost, nextPost, post, blogName, menuItems });
     let postPath = path.join(outFolder, post.filename) + '.html';
+    fs.ensureDirSync(path.dirname(postPath));
     fs.writeFileSync(postPath, rendered);
     console.log(`Published ${post.filename}`);
 }
